Changement des types de variable pour les raisons et U0

+++ Suite seules/Suite arithmétique.py
@@ -1,8 +1,8 @@
 print("Ce programme sert à générer une suite arithmétique fini avec une valeur \"U0\" et une raison \"q\" défini par l'utilisateur.")
 print()
-U=int(input("Valeur de \"U0\" : "))
+U=float(input("Valeur de \"U0\" : "))
 U0=U
-q=int(input("Valeur de la raison \"q\" : "))
+q=float(input("Valeur de la raison \"q\" : "))
 L=int(input("Longueur de la suite : "))
 P=0
 print("U0=",U0)
